fix(web): add validation to County field when editing IP comment address
