+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/add-ip-comment/add-ip-comment.router.js
@@ -9,9 +9,9 @@ import { Router as createRouter } from 'express';
 import * as controller from './add-ip-comment.controller.js';
 import {
 	validateCheckAddress,
-	validateInterestedPartyAddress,
 	validateInterestedPartyDetails
 } from './add-ip-comment.validators.js';
+import { validateInterestedPartyAddress } from '../common/validators.js';
 import { validateRedactionStatus } from '../../representations.validators.js';
 
 const router = createRouter({ mergeParams: true });
