+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/add-ip-comment/add-ip-comment.validators.js
@@ -1,16 +1,7 @@
 import { createValidator } from '@pins/express';
 import { body } from 'express-validator';
-import {
-	createTextInputOptionalValidator,
-	createTextInputValidator
-} from '#lib/validators/text-input-validator.js';
+import { createTextInputValidator } from '#lib/validators/text-input-validator.js';
 import { createEmailInputOptionalValidator } from '#lib/validators/email-input.validator.js';
-import {
-	createPostcodeValidator,
-	createAddressLine1Validator,
-	createAddressLine2Validator,
-	createTownValidator
-} from '#lib/validators/address.validator.js';
 import { textInputCharacterLimits } from '#appeals/appeal.constants.js';
 const maxLength = textInputCharacterLimits.defaultAddressInputLength;
 
@@ -34,16 +25,4 @@ export const validateInterestedPartyDetails = createValidator(
 		`Last name must be ${maxLength} characters or less`
 	),
 	createEmailInputOptionalValidator('emailAddress')
-);
-
-export const validateInterestedPartyAddress = createValidator(
-	createAddressLine1Validator(),
-	createAddressLine2Validator(),
-	createTownValidator(),
-	createTextInputOptionalValidator(
-		'county',
-		maxLength,
-		`County must be ${maxLength} characters or less`
-	),
-	createPostcodeValidator()
 );
