+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/edit-ip-comment/edit-ip-comment.router.js
@@ -1,13 +1,8 @@
 import { asyncHandler } from '@pins/express';
 import { Router as createRouter } from 'express';
-import {
-	createPostcodeValidator,
-	createAddressLine1Validator,
-	createAddressLine2Validator,
-	createTownValidator
-} from '#lib/validators/address.validator.js';
 import { createYesNoRadioValidator } from '#lib/validators/radio.validator.js';
 import * as controller from './edit-ip-comment.controller.js';
+import { validateInterestedPartyAddress } from '../common/validators.js';
 import { saveBodyToSession } from '#lib/middleware/save-body-to-session.js';
 
 const router = createRouter({ mergeParams: true });
@@ -16,10 +11,7 @@ router
 	.route('/address')
 	.get(asyncHandler(controller.renderEditAddress))
 	.post(
-		createAddressLine1Validator(),
-		createAddressLine2Validator(),
-		createTownValidator(),
-		createPostcodeValidator(),
+		validateInterestedPartyAddress,
 		saveBodyToSession('editIpComment'),
 		asyncHandler(controller.postEditAddress)
 	);
